Added these Inventory Operations:
-Add(Pick up)
-Drop(Remove)
-Move(from one slot to another)
-Swap(switch position of two items)
-Merge(moving item to another slot with same item)

+++ Channels/Event/SendDropPickUpRequestEvent.cs
@@ -33,9 +33,8 @@ namespace NineToFive.Event {
             Drop drop = user.Field.LifePools[EntityType.Drop][_objectId] as Drop;
             if (drop == null) return;
 
-            InventoryType inventoryType = ItemConstants.GetInventoryType((int) drop.TemplateId);
+            InventoryType inventoryType = ItemConstants.GetInventoryType(drop.TemplateId);
             Inventory inventory = user.Inventories[inventoryType];
-            
             if (drop.Id == 0) {
                 
             } else {
